Multiple fixes after erlang update

+++ scripts/setup-skyport-dev.sh
@@ -14,12 +14,11 @@ export SWM_VERSION_DIR=${SWM_ROOT}
 
 echo
 echo "=============== SETUP CLUSTER HEAD NODE ================"
-export SWM_API_PORT=10011
 ${ROOT_DIR}/scripts/setup-swm-core.py -x -t -v $SWM_VERSION -p $SWM_ROOT -s $SWM_SPOOL -c $CONFIG_BASE -d cluster -u $USER
 EXIT_CODE=$?
 if [ "$EXIT_CODE" != "0" ]; then
-  echo "Cluster setup command failed with code $EXIT_CODE"
-  exit $EXIT_CODE
+    echo "Cluster setup command failed with code $EXIT_CODE"
+    exit $EXIT_CODE
 fi
 
 echo
@@ -28,21 +27,21 @@ echo "=============== ADD EXTRA CONFIG ================"
 function wait_swm() {
   attempt=0
   EXIT_CODE=1
-  DIV=$1
-  COND=$2
+  CONDITION=$1
+  echo "Wait for Sky Port Daemon"
   while [ $attempt -lt 10 ]; do
     echo "Ping attempt $attempt"
-    ${ROOT_DIR}/scripts/run-in-shell.sh ${DIV} -p
-    if [ "$COND" == "started" ]; then
-      if [ "$?" == "0" ]; then
-        EXIT_CODE=0
-        break
-      fi
-    else
-      if [ "$?" != "0" ]; then
-        EXIT_CODE=0
-        break
-      fi
+    ${ROOT_DIR}/scripts/ping.erl localhost $SWM_API_PORT
+    if [ "$CONDITION" == "started" ]; then
+        if [ "$?" == "0" ]; then
+          EXIT_CODE=0
+          break
+        fi
+      else 
+        if [ "$?" != "0" ]; then
+          EXIT_CODE=0
+          break
+        fi
     fi
     (( attempt++ ))
     sleep 2
@@ -82,9 +81,10 @@ function import_config() {
   fi
 }
 
-# TODO: create sym links to porter and other scripts in /opt/swm from sources
-echo "Create symlink ~/.swm/spool --> /opt/swm/spool"
-ln -s /opt/swm/spool ~/.swm/spool
+echo "Ensure symlink ~/.swm/spool --> /opt/swm/spool"
+if [ ! -L "~/.swm/spool" ]; then
+  ln -s /opt/swm/spool ~/.swm/spool
+fi
 
 echo
 echo "Stop swm"
